requirements.txt
streamlit==1.46.0
streamlit-extras==0.7.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41